TC9 kernel (#1)
* TC9 kernel

* TC9 kernel

+++ Common/DtaDiskType.cpp
@@ -30,7 +30,7 @@ DtaDiskType::~DtaDiskType() {}
 #endif
 void DtaDiskType::init(const char * devref) {}
 uint8_t	DtaDiskType::sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-	void * buffer, uint16_t bufferlen) {
+	void * buffer, uint32_t bufferlen) {
 	return 10;
 }
 void DtaDiskType::identify(OPAL_DiskInfo& disk_info) {}
