+++ linux/DtaDevLinuxDrive.h
@@ -41,8 +41,7 @@ public:
      * @param buffer input/output buffer
      * @param bufferlen length of the input/output buffer
      */
-    virtual uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-            void * buffer, uint16_t bufferlen) = 0;
+    virtual uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,void * buffer, uint32_t bufferlen) = 0;
     /** Routine to send an identify to the device */
     virtual void identify(OPAL_DiskInfo& disk_info) = 0;
 };
